HOTFIX Cantidad recetas Int -> Float

+++ panqayuda/recetas/models.py
@@ -19,7 +19,7 @@ class Receta(models.Model):
 
     codigo = models.CharField(max_length=100, null=True, blank=False)
     nombre = models.CharField(max_length=100, null=True, blank=False)
-    cantidad = models.IntegerField(null=True, blank=False, validators=[MinValueValidator(1, "Debes seleccionar un número entero mayor a 0.")])
+    cantidad = models.FloatField(null=True, blank=False, validators=[MinValueValidator(1, "Debes seleccionar un número entero mayor a 0.")])
     duration = models.DurationField(null=True, blank=False)
     material = models.ManyToManyField('materiales.Material', through = 'RelacionRecetaMaterial')
     created_at = models.DateTimeField(default=timezone.now)
